+++ tests/unit/rules/prefer-caret-version-dependenciesTest.js
@@ -1,0 +1,45 @@
+"use strict";
+
+const should = require("should");
+const requireHelper = require("../../require_helper");
+const lint = requireHelper("rules/prefer-caret-version-dependencies").lint;
+
+describe("prefer-caret-version-dependencies Unit Tests", function() {
+  context("when package.json has node with an invalid value", function() {
+    it("LintIssue object should be returned", function() {
+      const packageJsonData = {
+        dependencies: {
+          "npm-package-json-lint": "~1.0.0"
+        }
+      };
+      const response = lint(packageJsonData);
+
+      response.lintId.should.equal("prefer-caret-version-dependencies");
+      response.lintType.should.equal("error");
+      response.node.should.equal("dependencies");
+      response.lintMessage.should.equal("You are using an invalid version range. Please use ^.");
+    });
+  });
+
+  context("when package.json has node with a valid value", function() {
+    it("LintIssue object should be returned", function() {
+      const packageJsonData = {
+        dependencies: {
+          "gulp-npm-package-json-lint": "^1.0.0"
+        }
+      };
+      const response = lint(packageJsonData);
+
+      response.should.be.true();
+    });
+  });
+
+  context("when package.json does not have node", function() {
+    it("true should be returned", function() {
+      const packageJsonData = {};
+      const response = lint(packageJsonData);
+
+      response.should.be.true();
+    });
+  });
+});